Fixed build_database when config has a space
In cases where config has a space in it (e.g. "Blarg - Debug"), the
output folder would also have a space in it. This broke
build_database_args and build_database_single so iccarm.exe would get
called with malformed arguments:
    [..., "-D", "Whatever I have defined", "-no-cse", ...] instead of
    [..., "-D", "Whatever I have defined", "--no-cse", ...]

The fix was to change the regex used in build_database args to require
a non-space character after the "-" since iccarm doesn't seem to
support "-" as an arg to indicate to read from stdin.

Also, the new regex doesn't require whitespace before the path, so no
fixup is required afterward to handle the first arg.

Additionally, arg_replacer has been replaced with a simple constant
string.